+++ sdk/ledgerTransaction/index.js
@@ -1,0 +1,7 @@
+const ledgerTransaction = require('./ledgerTransaction.js');
+
+exports.create = ledgerTransaction.create;
+exports.get = ledgerTransaction.get;
+exports.query = ledgerTransaction.query;
+exports.page = ledgerTransaction.page;
+exports.LedgerTransaction = ledgerTransaction.LedgerTransaction;